requirements.txt
Flask==2.3.3
Flask-Login==0.6.2
Flask-PyMongo==2.3.0
pymongo==4.6.0
Werkzeug==2.3.7
python-dotenv==1.0.0
email-validator==2.1.0
bcrypt==4.1.2
